Many changes

+++ src/main/java/com/fatwire/gst/web/servlet/profiling/servlet/jmx/ResponseTimeStatisticMBean.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2006 Dolf Dijkstra. All Rights Reserved.
+ * Copyright (C) 2006 Dolf Dijkstra
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *    http://www.apache.org/licenses/LICENSE-2.0
+ *         http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,10 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package com.fatwire.gst.web.servlet.profiling.servlet.jmx;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 
 public interface ResponseTimeStatisticMBean {
 
@@ -24,13 +23,13 @@ public interface ResponseTimeStatisticMBean {
      * 
      * @return the number of values in this statistic
      */
-    int getCount();
+    long getCount();
 
     /**
      * 
      * @return the total time of the values in this statistic
      */
-    BigDecimal getTotalTime();
+    BigInteger getTotalTime();
 
     /**
      * 
@@ -63,6 +62,7 @@ public interface ResponseTimeStatisticMBean {
      */
 
     public long getWaitCount();
+
     /**
      * 
      * @return the total time a thread has spend in system mode
@@ -70,6 +70,6 @@ public interface ResponseTimeStatisticMBean {
      * @see java.lang.management.ThreadMXBean#getCurrentThreadUserTime()
      */
 
-    public BigDecimal getTotalSystemTime();
+    public BigInteger getTotalSystemTime();
 
 }
